Split routes into separate controllers; refactor controller to remove configuration dependency; minor tidy-up in views
