+++ app/controllers/ApiController.scala
@@ -4,24 +4,14 @@ import model.CheckQuery
 import play.api.libs.json.{JsValue, Json}
 import play.api.mvc._
 import services._
-import play.api.Configuration
 import rules.RuleResource
 
 import scala.concurrent.{ExecutionContext, Future}
 
 /**
- * This controller creates an `Action` to handle HTTP requests to the
- * application's home page.
+ * The controller that handles API requests.
  */
-class ApiController(cc: ControllerComponents, validatorPool: ValidatorPool, ruleResource: RuleResource, configuration: Configuration)(implicit ec: ExecutionContext)  extends AbstractController(cc) {
-  def index() = Action { implicit request: Request[AnyContent] =>
-    Ok(views.html.index())
-  }
-
-  def healthcheck() = Action { implicit request: Request[AnyContent] =>
-    Ok("""{ "healthy" : "true" }""")
-  }
-
+class ApiController(cc: ControllerComponents, validatorPool: ValidatorPool, ruleResource: RuleResource)(implicit ec: ExecutionContext)  extends AbstractController(cc) {
   def check: Action[JsValue] = Action.async(parse.json) { request =>
     request.body.validate[CheckQuery].asEither match {
       case Right(checkQuery) =>
@@ -33,30 +23,6 @@ class ApiController(cc: ControllerComponents, validatorPool: ValidatorPool, rule
           Ok(json)
         })
       case Left(error) => Future.successful(BadRequest(s"Invalid request: $error"))
-    }
-  }
-
-  def refresh = Action.async { implicit request: Request[AnyContent] =>
-    for {
-      (rulesByCategory, ruleErrors) <- ruleResource.fetchRulesByCategory()
-      errorsByCategory <- Future.sequence(
-        rulesByCategory.map { case (category, rules) => {
-          validatorPool.updateConfig(category.name, ValidatorConfig(rules))
-        }}.toList
-      )
-      rules <- validatorPool.getCurrentRules
-    } yield {
-      val sheetId = configuration.getOptional[String]("typerighter.sheetId").orNull
-      val errors = errorsByCategory.flatten ::: ruleErrors
-      val rulesIngested = rulesByCategory.map { _._2.size }.sum
-      Ok(views.html.rules(sheetId, rules, Some(rulesIngested), errors))
-    }
-  }
-
-  def rules = Action.async { implicit request: Request[AnyContent] =>
-    val sheetId = configuration.getOptional[String]("typerighter.sheetId").getOrElse("No sheet id configured")
-    validatorPool.getCurrentRules.map { rules =>
-      Ok(views.html.rules(sheetId, rules))
     }
   }
 }
