+++ app/views/rules.scala.html
@@ -3,7 +3,7 @@
 
 @(sheetId: String, rules: scala.List[Rule], rulesIngested: Option[Int] = None, errors: List[String] = List.empty)(implicit req: RequestHeader)
 @main(title = "Typerighter", breadcrumbsList = List("Rules")) {
-    @form(CSRF(routes.ApiController.refresh())) {
+    @form(CSRF(routes.RulesController.refresh())) {
         @CSRF.formField
         <h1 class="align-middle">
             Current rules
@@ -28,7 +28,6 @@
                         <tr>
                             <th scope="col">Id</th>
                             <th scope="col">Colour</th>
-                            @*<th scope="col">Rule</th>*@
                             <th scope="col">Name</th>
                         </tr>
                     </thead>
@@ -42,7 +41,6 @@
                                     #@category.colour
                                 </span>
                             </td>
-                            @*<td>@if(rule.pattern.nonEmpty) { @rule.pattern } else { No&nbsp;rule }</td>*@
                             <td>@category.name</td>
                         </tr>
                     }
@@ -57,7 +55,6 @@
                     <tr>
                         <th scope="col">Category</th>
                         <th scope="col">Colour</th>
-                        @*<th scope="col">Rule</th>*@
                         <th scope="col">Description</th>
                     </tr>
                 </thead>
@@ -70,7 +67,6 @@
                                 #@rule.category.colour
                             </span>
                         </td>
-                        @*<td>@if(rule.pattern.nonEmpty) { @rule.pattern } else { No&nbsp;rule }</td>*@
                         <td>@if(rule.description.nonEmpty) {
                             @rule.description
                         } else { No&nbsp;description }</td>
